Promote vector length to long
Vector length on newer machines can easily exceed 31 bits, so we now
promote length from int to long.

+++ inc/bench.h
@@ -25,7 +25,7 @@ struct thread_args {
 
     /* Input */
     int tid;
-    int vlen;
+    long vlen;
 
     /* Thread management */
     pthread_mutex_t *mutex;
